+++ packages/perseus/src/widgets/interactive-graphs/graphs/point.tsx
@@ -1,5 +1,6 @@
 import * as React from "react";
 
+import {REMOVE_BUTTON_ID} from "../mafs-graph";
 import {actions} from "../reducer/interactive-graph-action";
 import useGraphConfig from "../reducer/use-graph-config";
 
@@ -80,12 +81,18 @@ export function UnlimitedPointGraph(props: PointGraphProps) {
                     onMove={(destination) =>
                         dispatch(actions.pointGraph.movePoint(i, destination))
                     }
-                    onFocusChange={(isFocused) => {
+                    onFocusChange={(event, isFocused) => {
                         if (isFocused) {
                             dispatch(actions.pointGraph.focusPoint(i));
                         } else {
+                            if (event.relatedTarget?.id === REMOVE_BUTTON_ID) {
+                                return;
+                            }
                             dispatch(actions.pointGraph.blurPoint());
                         }
+                    }}
+                    onClick={() => {
+                        dispatch(actions.pointGraph.clickPoint(i));
                     }}
                 />
             ))}