+++ packages/perseus/src/widgets/interactive-graphs/interactive-graph-question-builder.ts
@@ -7,6 +7,7 @@ import type {
     LockedFigureFillType,
     LockedFunctionType,
     LockedLabelType,
+    LockedLineStyle,
     LockedLineType,
     LockedPointType,
     LockedPolygonType,
@@ -16,10 +17,12 @@ import type {
 } from "../../perseus-types";
 import type {Interval, vec} from "mafs";
 
-export type LockedFunctionOptions = Omit<
-    Partial<LockedFunctionType>,
-    "type" | "equation"
->;
+export type LockedFunctionOptions = {
+    color?: LockedFigureColor;
+    strokeStyle?: LockedLineStyle;
+    directionalAxis?: "x" | "y";
+    domain?: Interval;
+};
 
 export function interactiveGraphQuestionBuilder(): InteractiveGraphQuestionBuilder {
     return new InteractiveGraphQuestionBuilder();
@@ -180,6 +183,11 @@ class InteractiveGraphQuestionBuilder {
         return this;
     }
 
+    withNoInteractiveFigure() {
+        this.interactiveFigureConfig = new NoInteractiveFigureConfig();
+        return this;
+    }
+
     withLinear(options?: {
         coords?: CollinearTuple;
         startCoords?: CollinearTuple;
@@ -296,7 +304,7 @@ class InteractiveGraphQuestionBuilder {
         point2: vec.Vector2,
         options?: {
             kind?: "line" | "ray" | "segment";
-            lineStyle?: "solid" | "dashed";
+            lineStyle?: LockedLineStyle;
             color?: LockedFigureColor;
             filled?: [boolean, boolean];
             showPoint1?: boolean;
@@ -499,6 +507,16 @@ class SegmentGraphConfig implements InteractiveFigureConfig {
             numSegments: this.numSegments,
             startCoords: this.startCoords,
         };
+    }
+}
+
+class NoInteractiveFigureConfig implements InteractiveFigureConfig {
+    correct(): PerseusGraphType {
+        return {type: "none"};
+    }
+
+    graph(): PerseusGraphType {
+        return {type: "none"};
     }
 }