+++ packages/perseus/src/widgets/interactive-graphs/stateful-mafs-graph.tsx
@@ -111,6 +111,7 @@ export const StatefulMafsGraph = React.forwardRef<
     const snapTo = graph.type === "polygon" ? graph.snapTo : null;
     const showAngles = graph.type === "polygon" ? graph.showAngles : null;
     const showSides = graph.type === "polygon" ? graph.showSides : null;
+    const startCoords = "startCoords" in graph ? graph.startCoords : null;
 
     const originalPropsRef = useRef(props);
     const latestPropsRef = useLatestRef(props);
@@ -131,7 +132,7 @@ export const StatefulMafsGraph = React.forwardRef<
         showAngles,
         showSides,
         latestPropsRef,
-        graph.startCoords,
+        startCoords,
     ]);
 
     // If the graph is static, it always displays the correct answer. This is